move some funcs and rename for new figure orderings
